Squashed commit of the following:
commit 066df319364aa4ad27c172e44d3d233d4df3aed4
Author: Mert Emir <48881002+mertemr@users.noreply.github.com>
Date:   Sat Jul 13 11:49:19 2024 +0000

    chore: auto detection for device is cuda or cpu

commit d4ac19b9c541132e6dd2dadf2cf5c2ef0b3ddc3b
Author: Mert Emir <48881002+mertemr@users.noreply.github.com>
Date:   Thu Jul 11 17:30:49 2024 +0300

    refactor: Update MongoDB find_one method to accept a Collection object

commit 8ed4c99729e052dcdc8fc551b8f52d728712386f
Author: Mert Emir <48881002+mertemr@users.noreply.github.com>
Date:   Thu Jul 11 17:23:38 2024 +0300

    chore: Add device argument to main.py for specifying the device to use for running the model

commit 722ba57e10e15f86995783795f8712021433d38b
Author: Mert Emir <48881002+mertemr@users.noreply.github.com>
Date:   Thu Jul 11 17:05:32 2024 +0300

    chore: Update main.py

commit 62ea19ff0937e4bc3881bb7f9549af13ffc8636c
Author: Mert Emir <48881002+mertemr@users.noreply.github.com>
Date:   Thu Jul 11 17:05:09 2024 +0300

    refactor: Improve response time calculation in find_closest_match function

commit eaa59a07f322c10309a3c848833f1f9ab35effda
Author: Mert Emir <48881002+mertemr@users.noreply.github.com>
Date:   Thu Jul 11 17:04:34 2024 +0300

    chore: Remove multiple code files related to Sentence_Sim module

commit f412ee41752f0b824250713c4990b928dea81d74
Author: Mert Emir <48881002+mertemr@users.noreply.github.com>
Date:   Thu Jul 11 16:19:27 2024 +0300

    Modules defined with 'as' instead of variable

commit f14f6c961fa1bb77f04d6b03bf74a030f00238b8
Author: Mert Emir <48881002+mertemr@users.noreply.github.com>
Date:   Thu Jul 11 16:17:58 2024 +0300

    chore: Refactor find_special and find_closest_match functions for improved readability and performance
