+++ utils/special_name.py
@@ -1,34 +1,31 @@
+import re
+from difflib import SequenceMatcher
 
-import difflib
-import time
+from utils.response_time import ShowResponseTime
 
-def find_special(SIM, text, special_list):
-    sentences = text.split(".")
-    for i in range(len(sentences)):
-        sentences[i] = sentences[i].strip().lower()
-
+def find_special(SIM, text: str, special_list: dict) -> dict[str, tuple[str, str, float]]:
+    sentences = [i.strip() for i in re.split(r'[.!?]', text.lower())]
     special_keys = list(special_list.keys())
-    most_dict = SIM.find_most_similar_sentences(special_keys, sentences)
-
-    for i in range(len(most_dict)):
-        key = most_dict[i][0]
+    
+    most_similar_sentences: dict[str, str] = SIM.find_most_similar_sentences(special_keys, sentences)
+    
+    for i, (key, sentence, similarity) in enumerate(most_similar_sentences):
         if key in special_list:
-            most_dict[i] = (special_list[key], most_dict[i][1], most_dict[i][2])
+            most_similar_sentences[i] = (special_list[key], sentence, similarity)
 
-    return most_dict
+    return most_similar_sentences
 
-def find_closest_match(prompt, special_list):
-    start_time = time.time()
-    prompt = prompt.lower()
+def find_closest_match(prompt: str, special_list: dict[str, str]) -> str:
+    prompt = prompt.lower().strip()
+    
     highest_ratio = 0
     closest_match = None
     
-    for key, value in special_list.items():
-        seq = difflib.SequenceMatcher(None, prompt, key.lower().strip())
-        ratio = seq.ratio()
-        
-        if ratio > highest_ratio:
-            highest_ratio = ratio
-            closest_match = value
-    print(f"Find Closest Response Time: {time.time()-start_time:.5f} seconds")
+    with ShowResponseTime("Find Closest"):
+        for key, value in special_list.items():
+            seq = SequenceMatcher(None, prompt, key.lower().strip())
+            if (ratio := seq.ratio()) > highest_ratio:
+                highest_ratio = ratio
+                closest_match = value
+    
     return closest_match
